refactor: create Flask app factory
